+++ domain/src/main/java/fr/insee/rem/domain/dtos/Context.java
@@ -1,0 +1,5 @@
+package fr.insee.rem.domain.dtos;
+
+public enum Context {
+    BUSINESS, HOUSEHOLD
+}