+++ domain/src/main/java/fr/insee/rem/domain/dtos/TypeUnit.java
@@ -1,5 +1,0 @@
-package fr.insee.rem.domain.dtos;
-
-public enum TypeUnit {
-    ENTERPRISE, HOUSEHOLD
-}